% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailPred.R
\name{mailPred}
\alias{mailPred}
\alias{mailPreds}
\title{Mail predict test for opencpu}
\usage{
mailPreds(id)
}
\arguments{
\item{id}{The row of mial data}
}
\description{
Mail predict test for opencpu.
}
\author{
Xu Jing
}
